requirements.txt
psycopg2-binary>=2.9,<2.10
python-dotenv>=1.0,<1.1
sqlalchemy>=2.0,<2.1
factory-boy>=3.3,<3.4